% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{check_config_file_exists}
\alias{check_config_file_exists}
\title{Checks if config file is present}
\usage{
check_config_file_exists()
}
\value{
list containing latest config file settings and a logical
 indicating if there is a config file.
}
\description{
Attempts to load in config file and check suitable config
 can be read via the function \code{\link{get_latest_settings}()}. Run
 when package is loaded.
}
