% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{check_for_config_file}
\alias{check_for_config_file}
\title{Check for grapho config file}
\usage{
check_for_config_file(config_file_location = NULL)
}
\arguments{
\item{config_file_location}{Location of config file.
If no location passed then system default will be used.}
}
\description{
Creates config file is it is not present
}
