% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{check_if_folder_is_writable}
\alias{check_if_folder_is_writable}
\title{Find out if folder can be written to}
\usage{
check_if_folder_is_writable(folder = NULL)
}
\arguments{
\item{folder}{Folder to try and write to}
}
\value{
TRUE if folder is writable and false if folder cannot be written to
}
\description{
Tries to write a file into a folder.
}
