% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_filename}
\alias{create_filename}
\title{Create filename}
\usage{
create_filename(filetype)
}
\arguments{
\item{filetype}{Type of file (logfile, plot, etc.) to be included in the
returned filename string.}
}
\description{
Used by
 Grapho when creating filenames and
returns the a filename string which includes the datetime, session ID,
user ID and filetype.
}
