% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_log_file}
\alias{create_log_file}
\title{create_log_file}
\usage{
create_log_file(return_location = FALSE, show_messages = TRUE)
}
\arguments{
\item{return_location}{If TRUE then location of log file is returned}

\item{show_messages}{If TRUE then message output displayed}
}
\description{
Run
 when Grapho is loaded. The log file is
created and the user is shown the location of the Grapho log file.
Grapho folder.
}
