% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorders.R
\name{error_recorder}
\alias{error_recorder}
\title{Records code errors.
The error recorder will write any errors to the log file. Commands
resulting in errors are not passed to the expression recorder, so we
also collect the most recent command from history.}
\usage{
error_recorder(error = geterrmessage())
}
\arguments{
\item{error}{error message}
}
\description{
Records code errors.
The error recorder will write any errors to the log file. Commands
resulting in errors are not passed to the expression recorder, so we
also collect the most recent command from history.
}
