% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_scribe.R
\name{error_scribe}
\alias{error_scribe}
\title{Error scribe}
\usage{
error_scribe(error = geterrmessage())
}
\arguments{
\item{error}{Error message. Default value is the output of
 \code{\link[base]{geterrmessage()}}
which returns the current error message.}
}
\description{
Writes to a file an error message and the most recent line sent
to the R console.
}
\details{
This function is called by the .zzz.R file run when the Grapho
package is first loaded into R. The function
 \code{\link{start_error_scribe()}} sets
\code{\link{error_scribe()}} as the error handling function.
}
