% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorders.R
\name{expression_recorder}
\alias{expression_recorder}
\title{Records expressions evaluated in the console.
The expression recorder is called when a top-level call is made.
 For instance, when assigning a variable. The call and contents
of plot devices written out to the Grapho directory and log
file set in the environment variables
GRAPHO_LOG_FILE and GRAPHO_FOLDER.}
\usage{
expression_recorder(top_level_expr, value, ok, visible)
}
\arguments{
\item{top_level_expr}{Top level expressions sent to the R console}

\item{value}{List of arguments passed in top level expression}

\item{ok}{Boolean indicating if call is ok}

\item{visible}{Boolean indicating if visible}
}
\description{
Records expressions evaluated in the console.
The expression recorder is called when a top-level call is made.
 For instance, when assigning a variable. The call and contents
of plot devices written out to the Grapho directory and log
file set in the environment variables
GRAPHO_LOG_FILE and GRAPHO_FOLDER.
}
