% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_scribe.R
\name{expression_scribe}
\alias{expression_scribe}
\title{Expression scribe}
\usage{
expression_scribe(top_level_expr, value, ok, visible)
}
\arguments{
\item{top_level_expr}{Top level expressions sent to the R console}

\item{value}{List of arguments passed in top level expression}

\item{ok}{Boolean indicating if call is ok}

\item{visible}{Boolean indicating if visible}
}
\value{
Logical value TRUE to indicate the callback is complete
}
\description{
The expression scribe is called when a top-level call is made. For instance,
when assigning a variable. The call and contents of plot devices written out
to the Grapho directory and log file set in the environment variables
GRAPHO_LOG_FILE and GRAPHO_FOLDER.
}
\details{
The \code{\link{start_expression_scribe()}}
 function is run when the Grapho
package is loaded and sets \code{\link{expression_scribe()}} as
a task callback function - the
 \code{\link{expression_scribe()}} is run after every
successful R console command. Errors which prevent a console command
finishing are recorded by the \code{\link{error_scribe()}} function.
}
