% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{get_config_file_location}
\alias{get_config_file_location}
\title{Retrieves the location of the config file}
\usage{
get_config_file_location()
}
\value{
String config file location
}
\description{
Gets the config file location
}
