% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_dependencies.R
\name{get_function_dependencies}
\alias{get_function_dependencies}
\title{Tries to identify package dependencies}
\usage{
get_function_dependencies()
}
\value{
Dataframe containing your package
}
\description{
\code{\link{get_function_dependencies()}} helps the user to identify the
packages used grapho archive.
}
\details{
Running \code{\link{get_function_dependencies()}} will read through
your Grapho archive to identify the packages required in each session. If you
do not have the packages installed then you
 will be prompted to do so. Finally,
the package for each function is identified using the
\code{{utils}\link{find()}} function.
}
