% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{get_latest_settings}
\alias{get_latest_settings}
\title{Retrieves the latest config settings}
\usage{
get_latest_settings()
}
\value{
Dataframe containing latest options
}
\description{
Loads in the current config file by
calling read_config_file() and then returns
the latest folder, plot_file_format and verbose option
}
