% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_information.R
\name{grapho_info}
\alias{grapho_info}
\title{Prints information about the grapho package}
\usage{
grapho_info()
}
\description{
Grapho is part of the What Aren't You
 Seeing (WAYS) project. The function print out
 information about the aims of the package and
 the authors.
}
