% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_information.R
\name{how_does_grapho_work}
\alias{how_does_grapho_work}
\title{An explanation of the inner workings of grapho}
\usage{
how_does_grapho_work()
}
\description{
Some may be interested in how grapho works. Understanding
 what grapho is doing is particularly important because grapho is
 storing your code activity in your grapho archive folder.
}
