% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_commands.R
\name{parse_commands}
\alias{parse_commands}
\title{Parses commands from Grapho archive}
\usage{
parse_commands()
}
\value{
Dataframe containing the token type, text, session ID, session time
and time for elements of commands.
}
\description{
Goes through your Grapho archive and
returns a summary of the R commands you have run.
}
\details{
Creates a Dataframe of your
Grapho archive using the \code{\link{parse_grapho_archive}()} command.
Commands are extracted from this Dataframe and then each command in turn is
passed to the \code{\link[utils]{getParseData}()} function which breaks down
the command into tokens such as SYMBOL_FUNCTION_ALL, SYMBOL, LEFT_ASSIGN, etc.
Tokens of type SYMBOL_FUNCTION_ALL, SYMBOL, LEFT_ASSIGN are extracted from
the output and labeled with the corresponding session id, sessions start time
and time.
}
