% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_grapho_archive.R
\name{parse_grapho_archive}
\alias{parse_grapho_archive}
\title{Parses Grapho archive}
\usage{
parse_grapho_archive()
}
\value{
Dataframe containing the time, type and record for commands, errors,
warnings and plots.
}
\description{
Reads the files
 in your Grapho archive and returns a
structured overview of your activities.
}
\details{
The \code{\link{parse_grapho_archive}()} function
 retrieves the locations of your
Grapho archive. For each logfile the function parses the logfile contents
retrieving the session ID, user ID, time, type and details of each event.
For each plot the function compares hashes of each plot file to remove
duplicates. The resulting dataframe offers a detailed overview
 of ones session.
}
