% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_archive.R
\name{prepare_archive}
\alias{prepare_archive}
\title{Prepares Grapho Archive}
\usage{
prepare_archive()
}
\description{
Creates a zip archive of
 all the files in the Grapho
folder and prints out a message informing the user where to submit the
zip archive.
}
\details{
Please do submit your Grapho data for the WAYS project at
the University of Warwick - we are looking at the workflow of R users and
would greatly appreciate your data submission.
}
