% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grapho.R
\name{print_random_log}
\alias{print_random_log}
\title{Prints the content of the current randomised log to the console}
\usage{
print_random_log()
}
\description{
Prints the content of the current randomised log to the console
}
