% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{print_welcome_message}
\alias{print_welcome_message}
\title{Welcome use to the grapho package}
\usage{
print_welcome_message(returning = TRUE)
}
\arguments{
\item{returning}{If TRUE message says welcome back}
}
\description{
Displays a welcome message introducing the user to grapho
}
