% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_archives}
\alias{read_archives}
\title{read_archives}
\usage{
read_archives()
}
\value{
Dataframe containing the time, type and record for commands, errors,
warnings and plots.
}
\description{
Reads the files
 in your Grapho archive and returns a
structured overview of your activities.
}
\details{
Go to each grapho archive location in
.grapho$config$grapho_archive$past and .grapho$config$grapho_archive$current 
load each of the log files and parse out the time, type, session and user 
from each. The function returns this data in a dataframe.
}
