% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{read_config_file}
\alias{read_config_file}
\title{Attempts to read the config file}
\usage{
read_config_file(print_location = FALSE)
}
\arguments{
\item{print_location}{If TRUE then config file location is also printed}
}
\value{
Dataframe containing the grapho configuration record
}
\description{
Config file is loaded in, the attempted load location is echoed
and a Dataframe containing the contents of the config file is returned
}
