% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grapho.R
\name{setup_grapho}
\alias{setup_grapho}
\title{Set the grapho configuration options}
\usage{
setup_grapho()
}
\description{
Provides an interface for setting up the grapho
configuration options. The current options are the grapho
arhive location and the image format used for plot files.
You can access the grapho configuration options by running 
.grapho$config at the console. The current options are saved 
in an RDS file at the location found at .grapho$grapho_state_file.
}
