% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_grapho_folder}
\alias{setup_grapho_folder}
\title{Setup Grapho folder}
\usage{
setup_grapho_folder(grapho_folder_location = NULL)
}
\arguments{
\item{grapho_folder_location}{specify folder location.
Used for testing purposes.}
}
\description{
Creates the Grapho folder, user and session IDs. Records the
folder location and ID information as global variables. Runs
when Grapho is loaded.
}
