% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{show_config}
\alias{show_config}
\title{Displays grapho config options in a user readable format.}
\usage{
show_config(return_dataframe = FALSE)
}
\arguments{
\item{return_dataframe}{If TRUE then a data frame containing your
current settings is returned}
}
\description{
Displays grapho config options in a user readable format.
}
