% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{start_expression_scribe}
\alias{start_expression_scribe}
\title{Starts expression scribe}
\usage{
start_expression_scribe()
}
\description{
Run when Grapho is loaded and
adds the expression scribe as a task callback. Expression scribe will
run when a command is sent to the R console.
}
