% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_grapho.R
\name{toggle_grapho}
\alias{toggle_grapho}
\title{Toggles Grapho recording}
\usage{
toggle_grapho()
}
\description{
Either stops or starts Grapho recording.
}
\details{
Displays either
 'Grapho logging disabled'
or 'Grapho logging enabled' to inform the user is the logging functionality
of the Grapho package has been enabled or disabled. The enabling or disabling
of Grapho will be recorded in the Grapho log file.

Grapho logging is set by the environment variable 'GRAPHO_LOGGING'. The
logging functionality in \code{\link{error_scribe}()} and
 \code{\link{expression_scribe}()} will only be
evaluated if GRAPHO_LOGGING is TRUE.
}
