% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{write_config}
\alias{write_config}
\title{Writes config setting to config file}
\usage{
write_config(folder = NULL, file_format = NULL, verbose = NULL)
}
\arguments{
\item{folder}{Folder to save grapho records to}

\item{file_format}{Preferred file format for graphics
files. JPG, PNG, SVG supported}

\item{verbose}{If TRUE then additional grapho messages are
displayed in the terminal}
}
\description{
Function is passed the folder, file_format and verbosity values
which are then written to the config file.
}
