% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{write_config_file}
\alias{write_config_file}
\title{Writes values to the grapho config file}
\usage{
write_config_file(option = NULL, value = NULL)
}
\arguments{
\item{option}{Config option being changed.
Either 'folder' or 'file_format' are accepted}

\item{value}{Value for new option e.g., folder location}
}
\description{
Config file is loaded and displayed
}
