% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{write_folder_location}
\alias{write_folder_location}
\title{Write folder location}
\usage{
write_folder_location(folder = NULL)
}
\arguments{
\item{folder}{Folder to save grapho records to}
}
\description{
Writes folder location to config file
}
