% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_plot}
\alias{write_plot}
\title{Saves the current plot to the grapho folder}
\usage{
write_plot(folder = NULL)
}
\arguments{
\item{folder}{folder to write plot into}
}
\description{
Creates a file containing the current
 plot into the grapho folder.
The filename follows the convention time-userID-sessionID.
The \code{\link[rstudioapi]{savePlotAsImage}()} is used if the function is
run within RStudio. If the function is not run in RStudio then
\code{\link[grDevices]{dev.copy}()} is used.
}
