% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{write_plot_file_format}
\alias{write_plot_file_format}
\title{Write verbostiy setting}
\usage{
write_plot_file_format(plot_file_format = NULL)
}
\arguments{
\item{plot_file_format}{file format of
either 'jpg', 'png', or 'svg'}
}
\description{
Writes verbosity option to config file
}
