% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{write_verbosity}
\alias{write_verbosity}
\title{Write verbostiy setting}
\usage{
write_verbosity(verbosity = NULL)
}
\arguments{
\item{verbosity}{verbosity setting of either TRUE or FALSE}
}
\description{
Writes verbosity option to config file
}
