% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_log_file}
\alias{create_log_file}
\title{create_log_file}
\usage{
create_log_file()
}
\description{
Run
 when Grapho is loaded. The log file is
created and the user is shown the location of the Grapho log file.
Grapho folder.
}
