% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_session_information}
\alias{log_session_information}
\title{Logs session information}
\usage{
log_session_information()
}
\description{
Run when
 Grapho is loaded and
records all of the R environment variables to a CSV file located in the
Grapho folder.
}
