% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{show_config}
\alias{show_config}
\title{Print out config file}
\usage{
show_config(console = FALSE)
}
\arguments{
\item{console}{If TRUE then forces the function to print out
to the console if run in RStudio}
}
\description{
Config file is loaded using the function
\code{\link{read_config_file}()} and displayed
}
