% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{start_error_scribe}
\alias{start_error_scribe}
\title{Starts error scribe}
\usage{
start_error_scribe()
}
\description{
\code{\link{start_error_scribe()}} is run
 when Grapho is loaded and
changes the error options to run the error scribe when an error occurs.
}
